+++ novel_downloader/core/requesters/common_requester/common_session.py
@@ -47,7 +47,7 @@ class CommonSession(BaseSession):
         self._site = site
         self._profile = profile
 
-    def get_book_info(self, book_id: str, wait_time: Optional[int] = None) -> str:
+    def get_book_info(self, book_id: str, wait_time: Optional[float] = None) -> str:
         """
         Fetch the raw HTML (or JSON) of the book info page.
 
@@ -75,7 +75,7 @@ class CommonSession(BaseSession):
         raise RuntimeError("Unexpected error: get_book_info failed without returning")
 
     def get_book_chapter(
-        self, book_id: str, chapter_id: str, wait_time: Optional[int] = None
+        self, book_id: str, chapter_id: str, wait_time: Optional[float] = None
     ) -> str:
         """
         Fetch the raw HTML (or JSON) of a single chapter.
